+++ package.json
@@ -13,7 +13,8 @@
     "eslint": "7.31.0",
     "eslint-config-htmlacademy": "4.0.2",
     "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12"
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0"
   },
   "engines": {
     "node": "12"
