+++ src/utils/create-component.js
@@ -1,0 +1,9 @@
+export const createComponent = (wrap, content, point) => {
+  if (point === 'start') {
+    wrap.prepend(content);
+  }
+
+  if (point === 'end') {
+    wrap.append(content);
+  }
+}
