+++ webpack.config.js
@@ -3,7 +3,7 @@ const path = require('path');
 module.exports = {
   entry: './src/main.js',
   output: {
-    path: path.resolve(__dirname, 'public'),
+    path: path.resolve(__dirname, './public'),
     filename: 'bundle.js'
   },
   devtool: 'source-map',
